pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dnsrobocert"
version = "3.26.1"
description = "A tool to manage your DNS-challenged TLS certificates"
license = "MIT"
keywords = [
    "dnsrobocert"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix", "Operating System :: POSIX",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Utilities"
]
authors = [
    { name = "Adrien Ferrand", email = "ferrand.ad@gmail.com" },
]
readme = "README.rst"

requires-python = ">=3.11"
dependencies = [
    "acme>=5.0,<6.0",
    "certbot>=5.0,<6.0",
    "dns-lexicon[full] @ git+https://github.com/jonmeacham/dns-lexicon.git@bc109536ac26c541de7f98318f49a8660d014431",
    "cryptography>=41.0,<47.0",
    "cffi>=1.15,<3.0",
    "lxml>=4.9,<7.0",
    "dnspython>=2.1.0,<3.0",
    "jsonschema>=4.0,<5.0",
    "pem>=21.0,<24.0",
    "pyopenssl>=23.0,<26.0",
    "pyyaml>=6.0,<7.0",
    "schedule>=1.1,<2.0",
    "coloredlogs>=15.0,<16.0",
    "colorama>=0.4,<1.0",
    "watchdog>=4.0,<5.0",
    "tldextract>=3.4,<6.0",
]

[project.urls]
repository = "https://github.com/adferrand/dnsrobocert"
homepage = "https://adferrand.github.io/dnsrobocert/"

[project.scripts]
dnsrobocert = "dnsrobocert.core.main:main"

[dependency-groups]
dev = [
    "black>=23.0,<26.0",
    "flake8>=6.0,<8.0",
    "flake8-pyproject>=1.2,<2.0",
    "mypy>=1.5,<2.0",
    "packaging>=21.0,<26.0",
    "pytest>=7.0,<9.0",
    "pytest-cov>=4.0,<8.0",
    "pytest-sugar>=0.9,<2.0",
    "isort>=5.10,<7.0",
    "types-PyYAML>=6.0,<7.0",
    "types-pyOpenSSL>=23.0,<25.0",
    "types-requests>=2.28,<3.0",
    "types-setuptools>=65.0,<81.0",
]
doc = [
    "sphinx>=8.0,<9.0",
    "piccolo-theme>=0.22,<1.0",
    "toml>=0.10,<1.0"
]

[tool.pytest.ini_options]
junit_family = "xunit2"
addopts = "-p no:legacypath"

[tool.mypy]
show_error_codes = true
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E501", "E231", "E221"]

[tool.tox.env_run_base]
runner = "uv-venv-lock-runner"
with_dev = true
commands = [[
    "pytest", "test", "--doctest-modules", "--junitxml=junit/test-results.xml",
]]

[tool.tox.env.cover]
commands = [[
    "pytest", "test", "--doctest-modules", "--junitxml=junit/test-results.xml",
    "--cov=dnsrobocert", "--cov-report=term-missing", "--cov-report=xml",
]]

[tool.tox.env.flake8]
commands = [[
    "flake8", "src", "test", "utils",
]]

[tool.tox.env.mypy]
commands = [[
    "mypy", "src",
]]

[tool.hatch.metadata]
allow-direct-references = true
